refactor(schedule): split view into focused components and composables
